article/update finish

+++ pages/article/editor/[id]/index.js
@@ -1,9 +1,9 @@
 import dynamic from 'next/dynamic';
 import Router from 'next/router';
 
-import { Button, message } from 'antd';
+import { Button, message, Modal } from 'antd';
 import { Layout } from '../../../../containers';
-import { deleteArticle, fetchArticle } from '../../../../api';
+import { deleteArticle, fetchArticle, putArticle } from '../../../../api';
 
 import Style from './index.style';
 
@@ -12,30 +12,45 @@ const DynamicEditorNossr = dynamic(
   { ssr: false }
 );
 
+const { confirm } = Modal;
+
 const ArticleEdit = ({ articleMap, id }) => {
   // 删除文章
-  const handleRemoveArticle = async () => {
-    try {
-      const { code } = await deleteArticle(id);
-      if(code === 0){
-        message.success('删除成功');
-        Router.push('/article');
-      }else{
-        message.error('删除失败');
-      }
-    } catch (error) {
-      console.warn(error);
-    }
+  const handleRemoveArticle = () => {
+    confirm({
+      title: '删除',
+      content: '确定删除此文章吗',
+      okText: 'Yes',
+      okType: 'danger',
+      cancelText: 'No',
+      onOk: async () => {
+        try {
+          const { code } = await deleteArticle(id);
+          if (code === 0) {
+            message.success('删除成功');
+            Router.push('/article');
+          } else {
+            message.error('删除失败');
+          }
+        } catch (error) {
+          console.warn(error);
+        }
+
+      },
+      onCancel() {
+        console.log('Cancel');
+      },
+    });
   };
 
-  // 保持文章方法
+  // 保存文章方法
   const handleSubmitEditorState = async (submitData) => {
     try {
-      const { code, data } = await postArticle(submitData);
-      if(code === 0){
+      const { code, data } = await putArticle(id, submitData);
+      if (code === 0) {
         message.success('保存成功');
         Router.push('/article');
-      }else {
+      } else {
         message.error(`保存失败! ${ data.msg }`);
       }
     } catch (error) {
